Vote and unvote - update admin dashboard

+++ app/controllers/votes_controller.rb
@@ -2,6 +2,9 @@ class VotesController < ApplicationController
   before_action :find_contestant
 
   def create
+    existing_vote = @contestant.votes.where(user: current_user).first
+    return if existing_vote
+
     vote = @contestant.votes.create!(user: current_user)
     # ...
     @contestant.my_vote = vote
@@ -10,6 +13,8 @@ class VotesController < ApplicationController
 
   def destroy
     existing_vote = @contestant.votes.where(user: current_user).first
+    return unless existing_vote
+
     existing_vote.destroy!
     # ...
     render partial: "activity/votes", locals: { contestant: @contestant }
